10.95 Policy Authorization in Controller

+++ app/Http/Controllers/JobController.php
@@ -4,13 +4,16 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\View\View;
-use App\Models\Job;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use App\Models\Job;
 
 class JobController extends Controller
 {
+    use AuthorizesRequests;
+
     /**
      * @desc Show all job listings
      * @route GET /jobs
@@ -82,6 +85,9 @@ class JobController extends Controller
      */
     public function edit(Job $job): View
     {
+        // Check if user is authorized to edit the job
+        $this->authorize('update', $job);
+
         return view('jobs.edit')->with('job', $job);
     }
 
@@ -94,6 +100,9 @@ class JobController extends Controller
      */
     public function update(Request $request, Job $job): RedirectResponse
     {
+        // Check if user is authorized to edit the job
+        $this->authorize('update', $job);
+
         $validatedData = $this->validatedData($request);
 
         // Check for image
@@ -121,6 +130,8 @@ class JobController extends Controller
      */
     public function destroy(Job $job): RedirectResponse
     {
+        // Check if user is authorized to delete the job
+        $this->authorize('delete', $job);
 
         if ($job->company_logo) {
             $this->deleteLogo($job->company_logo);
